add missing dependency and command
First, thank you so much for this. It was the only piece of advice I could find anywhere on how to get linux on my old transformer tablet.
I just did this whole process and had some issues, for which I've set the required commands i had to use in here for future reference. Basically just `autopoint` was missing as a dependency and grub's `autogen.sh` required `./bootstrap` to be executed first.